+++ addons/de/xrechnung/invoices_test.go
@@ -107,7 +107,7 @@ func testInvoiceStandard(t *testing.T) *bill.Invoice {
 		Ordering: &bill.Ordering{
 			Code: "1234567890",
 		},
-		Payment: &bill.Payment{
+		Payment: &bill.PaymentDetails{
 			Instructions: &pay.Instructions{
 				Key: "credit-transfer",
 				CreditTransfer: []*pay.CreditTransfer{