+++ bill/invoice.go
@@ -74,9 +74,9 @@ type Invoice struct {
 	// Ordering details including document references and buyer or seller parties.
 	Ordering *Ordering `json:"ordering,omitempty" jsonschema:"title=Ordering Details"`
 	// Information on when, how, and to whom the invoice should be paid.
-	Payment *Payment `json:"payment,omitempty" jsonschema:"title=Payment Details"`
+	Payment *PaymentDetails `json:"payment,omitempty" jsonschema:"title=Payment Details"`
 	// Specific details on delivery of the goods referenced in the invoice.
-	Delivery *Delivery `json:"delivery,omitempty" jsonschema:"title=Delivery Details"`
+	Delivery *DeliveryDetails `json:"delivery,omitempty" jsonschema:"title=Delivery Details"`
 
 	// Summary of all the invoice totals, including taxes (calculated).
 	Totals *Totals `json:"totals" jsonschema:"title=Totals" jsonschema_extras:"calculated=true"`