+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Changed
+
+- `bill`: renaming `Payment` to `PaymentDetails`, and `Delivery` to `DeliveryDetails`, to make room for new document types.
+
 ## [v0.210.0] - 2025-02-19
 
 ### Added